requirements.txt
absl-py==0.15.0
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
antlr4-python3-runtime==4.8
anyio==3.7.1
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.3
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
clang==5.0
clip==0.2.0
crcmod==1.7
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
datasets==2.13.0
decorator==5.1.1
dill==0.3.6
easydict==1.9
einops==0.3.0
exceptiongroup==1.2.0
filelock==3.12.2
flatbuffers==1.12
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==1.0.0
gast==0.4.0
google-auth==2.28.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz @ file:///opt/conda/conda-bld/python-graphviz_1660063183935/work
grpcio==1.62.0
h11==0.14.0
h5py==3.1.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.2
importlib-metadata==6.7.0
jmespath==0.10.0
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kornia==0.5.8
Levenshtein==0.23.0
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
modelscope==1.9.5
multidict==6.0.5
multiprocess==0.70.14
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
oss2==2.18.4
packaging==23.2
paddlepaddle-gpu==2.5.2
pandas==1.3.5
Pillow==9.4.0
platformdirs==4.0.0
protobuf==3.20.3
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
pyDeprecate==0.3.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-lightning==1.4.2
pytz==2024.1
PyWavelets==1.3.0
PyYAML==6.0.1
rapidfuzz==3.4.0
regex==2023.12.25
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.16.2
scipy==1.7.3
simplejson==3.19.2
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.0
tensorflow-estimator==2.15.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.6.0
torchvision==0.14.1
tqdm==4.66.2
transformers==4.19.2
typing-extensions==3.7.4.3
urllib3 @ file:///croot/urllib3_1673575502006/work
Werkzeug==2.2.3
wrapt==1.12.1
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.15.0
